Incluir proteção de rotas e organização de componentes em pastas
