+++ front/src/App.tsx
@@ -1,8 +1,8 @@
 import { Suspense } from "react";
-import { Loading } from "./components/Loading";
-import Menu from "./components/Menu";
-import LayoutRestaurante from "./components/LayoutRestaurante";
-import AppRoutes from "./components/AppRoutes";
+import { Loading } from "./components/shared/Loading";
+import Menu from "./components/shared/Menu";
+import LayoutRestaurante from "./components/Restaurante/LayoutRestaurante";
+import AppRoutes from "./routes/AppRoutes";
 import { useAuth } from "./contexts/AuthContext";
 
 function App() {
