+++ front/src/components/Auth/AuthFacebook.tsx
@@ -1,4 +1,4 @@
-import Button from "../Button";
+import Button from "../ui/Button";
 import { FaFacebook } from "react-icons/fa";
 
 export default function AuthFacebook() {
