+++ front/src/components/Restaurante/MenuRestaurante.tsx
@@ -8,7 +8,7 @@ import {
 } from "lucide-react";
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../../contexts/AuthContext";
 
 const MenuRestaurante = () => {
   const [isOpen, setIsOpen] = useState(false);
