+++ front/src/contexts/AuthContext.tsx
@@ -1,4 +1,3 @@
-// src/contexts/AuthContext.tsx
 import React, { createContext, useContext, useState, useEffect } from "react";
 import { decodeToken } from "../utils/decodeToken";
