+++ front/src/pages/Account.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
-import UserForm from "../components/UserAccount";
-import RestaurantForm from "../components/RestaurantAccount";
-import Button from "../components/Button";
+import UserForm from "../components/Usuario/UserAccount";
+import RestaurantForm from "../components/Restaurante/RestaurantAccount";
+import Button from "../components/ui/Button";
 import { useAuth } from "../contexts/AuthContext";
 
 export default function Account() {
