+++ front/src/pages/Intermediaria.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import Button from "../components/Button";
+import Button from "../components/ui/Button";
 
 export default function Intermediaria() {
   const navigate = useNavigate();
