+++ front/src/routes/ProtectedRoute.tsx
@@ -1,0 +1,8 @@
+import { Navigate, Outlet } from "react-router-dom";
+import { useAuth } from "../contexts/AuthContext";
+
+export default function ProtectedRoute() {
+  const { isAuthenticated } = useAuth();
+
+  return isAuthenticated ? <Outlet /> : <Navigate to="/auth" replace />;
+}
