+++ front/src/utils/decodeToken.tsx
@@ -5,7 +5,6 @@ export const decodeToken = (token: string): ITokenData | null => {
   if (!token) return null;
 
   try {
-    // console.log(jwtDecode<ITokenData>(token))
     return jwtDecode<ITokenData>(token);
   } catch (error) {
     console.error("Invalid token", error);
